Update setup scripts

+++ setup/deploy.sh
@@ -1,31 +1,31 @@
-#! /bin/sh
-
-# Prerequites:
-# - this script file should be in the same directory as the project folder
-#   - e.g. ~/jeeves-bot, ~/deploy-jeeves-bot.sh
-# - secrets directory and file in the same directory as this script,
-#   - e.g. ~/secrets/.env
+#! /bin/bash
 
 if ! [ $(id -u) = 0 ]; then
     echo "Please run as root"
     exit 1
 fi
+if ! [ -a "/home/ubuntu/secrets/.env" ]; then
+    echo "Missing secrets file /home/ubuntu/secrets/.env"
+    exit 1
+fi
+
+export NODE_ENV=production
 
 echo "Stopping application"
 systemctl stop jeeves-bot
 
 echo "Getting latest code"
-[ -d "jeeves-bot" ] && rm -rf jeeves-bot
-# only run systemd commands as root
+[ -d "/home/ubuntu/jeeves-bot" ] && rm -rf "/home/ubuntu/jeeves-bot"
+# do not run git as root
 sudo -u ubuntu git clone https://github.com/miljinx/jeeves-bot.git
 
 echo "Copying secrets into project directory"
-sudo -u ubuntu cp secrets/.env jeeves-bot/.env
+sudo -u ubuntu cp "/home/ubuntu/secrets/.env" "/home/ubuntu/jeeves-bot/.env"
 
 echo "Setting up node"
 cd jeeves-bot
-# disable husky setup
-sudo -u ubuntu npm set-script prepare ""
+# do not run npm as root
+sudo -u ubuntu npm set-script prepare "" # disable dev script
 sudo -u ubuntu npm install
 
 echo "Deploying commands"
